Separates 'name' from credentials in Bluemix connections.
It doesn't make a lot of sense for the configuration name
to be added to the credentials object. This was confusing some
users. So now, 'name' is separate in the object constructor, but
is currently optional.
The key 'name' in credentials is still supported, but a warning
message is printent when used.