Now raises exception for invalid names with swift protocol.
It seems that the swift protocol (version 1) does not support
underscores in the swift URL -- either in the 'name' of the
configuration or in the name of the object storage container!

This commit raises ValueError exceptions when underscores are found,
making it transparent to the user why the URL would fail, rather
than the generic 'invalid host name' error message.