Update README.md
Changes the documented way to `import ibmos2spark`.  Removes aliasing the package to `oss` in favor
of using `ibmos2spark`. No need to pollute the users's namespace with `oss`, which might conflict with something else. 
It won't be used often, so typing out 'ibmos2spark' in full a couple of times isn't a big deal.