Removes old swift protocol support.
This commit removes configuration of the older swift protocol
in favor of the new swift2d/stocator protocol.

All classes, however, still persist (softlayer, softlayer2d,
bluemix, bluemix2d). The '2d' classes are subclasses of their
parents, however. They produce a warning when used and will
be deprecated in the future.

An attempt was made to be minimally disruptive for
older code that uses the softlayer class by accepting
the older set of arguments. This will be quickly deprecated,
however.