Updating README in python subdirectory.
This update clearly shows how to configure Spark to
Object Storage connections for the different flavors offered
by IBM.

Some of the previous sample codes were slightly modified for clarity.

The titles of the sections were changed to match their titles in the
IBM Bluemix catalog of Object Storage services.

The order of the sample codes were arranged to match the order,
from left to right, of the Object Storage services listed in the
IBM Bluemix catalog.